refactor: split oversized *_widgets.dart files into part files
The four largest presentation files (home_widgets 576, profile_widgets
503, notifications_widgets 452, bookmarks_list_tile 436 lines) each bundled
one public widget plus many private helper widgets in a single file, making
them hard to navigate.

Split each into focused `part` files grouped by concern (e.g. home_welcome
/ home_stats / home_recent). Using `part`/`part of` keeps the helper
widgets library-private and leaves the public import surface unchanged, so
no consumer imports change. Pure reorganization — no behavior change.
