+++ lib/features/bookmarks/presentation/widgets/bookmarks_list_tile_actions.dart
@@ -1,0 +1,46 @@
+part of 'bookmarks_list_tile.dart';
+
+Future<bool> _confirmDelete(BuildContext context, String title) async {
+  final l10n = context.l10n;
+  return await showDialog<bool>(
+        context: context,
+        builder: (ctx) => AlertDialog(
+          title: Text(l10n.bookmarkDeleteDialogTitle),
+          content: Text(l10n.bookmarkDeleteDialogMessage(title)),
+          actions: [
+            AppButton(
+              label: l10n.commonCancel,
+              variant: AppButtonVariant.text,
+              onPressed: () => Navigator.of(ctx).pop(false),
+            ),
+            AppButton(
+              label: l10n.commonDelete,
+              variant: AppButtonVariant.tonal,
+              onPressed: () => Navigator.of(ctx).pop(true),
+            ),
+          ],
+        ),
+      ) ??
+      false;
+}
+
+Future<void> _showItemMenu(BuildContext context, Bookmark bookmark) async {
+  final l10n = context.l10n;
+  final result = await showModalBottomSheet<String>(
+    context: context,
+    builder: (sheetContext) => SafeArea(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          ListTile(
+            leading: const FaIcon(FontAwesomeIcons.shareNodes),
+            title: Text(l10n.commonShare),
+            onTap: () => Navigator.pop(sheetContext, 'share'),
+          ),
+        ],
+      ),
+    ),
+  );
+  if (result != 'share' || !context.mounted) return;
+  context.read<BookmarksListBloc>().add(BookmarksListShareRequested(bookmark));
+}
