+++ lib/features/profile/presentation/widgets/profile_about.dart
@@ -1,0 +1,76 @@
+part of 'profile_widgets.dart';
+
+class _AppInfoTile extends StatelessWidget {
+  const _AppInfoTile();
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<ProfileBloc, ProfileState>(
+      builder: (context, state) {
+        final info = state.packageInfo;
+        return ListTile(
+          leading: const FaIcon(FontAwesomeIcons.circleInfo),
+          title: Text(info?.appName ?? '—'),
+          subtitle: info == null
+              ? Text(context.l10n.commonLoading)
+              : Text(
+                  context.l10n.profileAppVersionBuild(
+                    info.version,
+                    info.buildNumber,
+                  ),
+                ),
+        );
+      },
+    );
+  }
+}
+
+class _SignOutButton extends StatelessWidget {
+  const _SignOutButton();
+
+  @override
+  Widget build(BuildContext context) {
+    final session = SessionScope.of(context);
+    return ListenableBuilder(
+      listenable: session,
+      builder: (context, _) {
+        final isLoading = session.isSigningOut;
+        return Padding(
+          padding: const EdgeInsets.symmetric(horizontal: AppSpacing.xl),
+          child: AppButton(
+            label: context.l10n.commonSignOut,
+            icon: FontAwesomeIcons.arrowRightFromBracket,
+            variant: AppButtonVariant.tonal,
+            expand: true,
+            isLoading: isLoading,
+            onPressed: isLoading ? null : () => _confirmSignOut(context),
+          ),
+        );
+      },
+    );
+  }
+
+  Future<void> _confirmSignOut(BuildContext context) async {
+    final l10n = context.l10n;
+    final confirmed = await showDialog<bool>(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        title: Text(l10n.commonSignOut),
+        content: Text(l10n.profileSignOutConfirmMessage),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.of(ctx).pop(false),
+            child: Text(l10n.commonCancel),
+          ),
+          TextButton(
+            onPressed: () => Navigator.of(ctx).pop(true),
+            child: Text(l10n.commonSignOut),
+          ),
+        ],
+      ),
+    );
+    if (confirmed == true && context.mounted) {
+      SessionScope.of(context).signOut();
+    }
+  }
+}
