+++ lib/features/profile/presentation/widgets/profile_header.dart
@@ -1,0 +1,111 @@
+part of 'profile_widgets.dart';
+
+class _ProfileHeader extends StatelessWidget {
+  const _ProfileHeader();
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final session = SessionScope.of(context);
+    return ListenableBuilder(
+      listenable: session,
+      builder: (context, _) {
+        final user = session.currentUser;
+        final username = user?.username ?? '';
+        final initial = username.isNotEmpty ? username[0].toUpperCase() : '?';
+        return Column(
+          children: [
+            CircleAvatar(
+              radius: 48,
+              backgroundColor: theme.colorScheme.primaryContainer,
+              child: Text(
+                initial,
+                style: theme.textTheme.headlineMedium?.copyWith(
+                  color: theme.colorScheme.onPrimaryContainer,
+                ),
+              ),
+            ).animateScale(),
+            const SizedBox(height: AppSpacing.md),
+            Text(
+              username,
+              style: theme.textTheme.titleLarge,
+            ).animateSlideDown(delay: 150.ms),
+            const SizedBox(height: AppSpacing.xs),
+            _CopyableId(id: user?.id ?? '').animateFadeIn(delay: 250.ms),
+          ],
+        );
+      },
+    );
+  }
+}
+
+class _CopyableId extends StatelessWidget {
+  const _CopyableId({required this.id});
+
+  final String id;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    if (id.isEmpty) return const SizedBox.shrink();
+    return InkWell(
+      borderRadius: BorderRadius.circular(AppRadius.sm),
+      onTap: () {
+        Clipboard.setData(ClipboardData(text: id));
+        context.read<ProfileBloc>().add(const ProfileUserIdCopied());
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text(context.l10n.profileUserIdCopied)),
+        );
+      },
+      child: Padding(
+        padding: const EdgeInsets.symmetric(
+          horizontal: AppSpacing.sm,
+          vertical: AppSpacing.xs,
+        ),
+        child: Row(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text(
+              id,
+              style: theme.textTheme.bodySmall?.copyWith(
+                color: theme.colorScheme.onSurfaceVariant,
+              ),
+            ),
+            const SizedBox(width: AppSpacing.xs),
+            FaIcon(
+              FontAwesomeIcons.copy,
+              size: 14,
+              color: theme.colorScheme.onSurfaceVariant,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class _SectionLabel extends StatelessWidget {
+  const _SectionLabel(this.text);
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(
+        AppSpacing.xl,
+        AppSpacing.xs,
+        AppSpacing.xl,
+        AppSpacing.sm,
+      ),
+      child: Text(
+        text.toUpperCase(),
+        style: theme.textTheme.labelSmall?.copyWith(
+          color: theme.colorScheme.primary,
+          letterSpacing: 0.8,
+        ),
+      ),
+    );
+  }
+}
